wonderai/backend/requirements.txt
# WonderAI Backend Requirements
# FastAPI with modern Python patterns and AI integrations

# Core Framework
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.22

# Database & ORM
sqlmodel==0.0.22
alembic==1.14.0
psycopg2-binary==2.9.10

# AI & ML Libraries
openai==1.58.1
langchain==0.3.9
langchain-openai==0.2.9
langchain-community==0.3.9
faiss-cpu==1.9.0
sentence-transformers==3.3.1
tiktoken==0.8.0

# HTTP Client & API Tools
httpx==0.27.2
aiohttp==3.11.10
requests==2.32.3

# Environment & Configuration
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1

# Security & Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.22

# WebSocket Support
websockets==14.1

# Logging & Monitoring
loguru==0.7.2
structlog==24.4.0

# Development Tools
pytest==8.3.4
pytest-asyncio==0.24.0
black==24.10.0
ruff==0.8.4

# Image Processing
Pillow==11.0.0

# Additional Utilities
redis==5.2.1
celery==5.4.0
rich==13.9.4
